airflow/requirements.txt
apache-airflow-providers-apache-spark==4.8.1
apache-airflow-providers-apache-kafka==1.5.0
kafka-python==2.0.2
pyspark==3.3.2
